requirements.txt
Django>=3.2.0
psycopg2-binary>=2.9.0
djangorestframework>=3.12.0
django-cors-headers>=3.11.0
djangorestframework-simplejwt>=4.8.0
dj-database-url==2.1.0